Add files via upload
These were some coding exercises to do in python, lab logs of how I did each activity was included and there is a many report for some given loan data for analysis. The activity were the exercises which needed to be completed. The train.csv is the data set used for the loan data analysis in week 3.